refactor: 페이지별 훅 및 타입 분리 / feat: 채팅방 생성 여부 체크 후 이동 기능 추가
