+++ src/app/page.tsx
@@ -1,8 +1,5 @@
 // 메인페이지
-
-import MainPage from "@/components/main/MainPage";
-
-
+import MainPage from "@/components/MainPage/";
 
 export default function Home() {
   return (
