+++ src/components/communityBoard/CommunityBoard-detail/Comment/index.tsx
@@ -38,19 +38,12 @@ export default function Comment() {
   return (
     <>
       <section>
-        {/* 하단 입력 & 북마크 버튼 */}
         <div className="w-full bg-[#FDFCF8]">
-          {/* 만든거 */}
-          {/* 댓글 리스트 */}
           <div className="space-y-6 px-5 pb-20 ">
-            {/* 개별 댓글 */}
             <div className="flex flex-col gap-4">
-              {/* ---------- */}
               {comments.map((comment) => (
                 <div key={comment.id} className="space-y-4">
-                  {/* 🟢 댓글 (기존 코드) */}
                   <div className="flex items-start space-x-3">
-                    {/* 프로필 이미지 */}
                     <div
                       className="w-12 h-12 rounded-3xl bg-center bg-cover bg-no-repeat bg-[#d3d3d3] flex-shrink-0"
                       style={{
@@ -58,7 +51,6 @@ export default function Comment() {
                       }}
                     ></div>
 
-                    {/* 댓글 내용 */}
                     <div className="flex-1">
                       <div className="flex items-center justify-between w-full">
                         <div className="flex items-center space-x-2">
@@ -66,9 +58,14 @@ export default function Comment() {
                             {comment.writeUserName}
                           </span>
                           <span className="text-gray-500 text-sm">
-                            {new Date(comment.createdAt).toLocaleDateString(
-                              "ko-KR"
-                            )}
+                            {new Date(comment.createdAt)
+                              .toLocaleDateString("ko-KR", {
+                                year: "numeric",
+                                month: "2-digit",
+                                day: "2-digit",
+                              })
+                              .replace(/-/g, ".")}{" "}
+                            {/* 하이픈을 점으로 변경 */}
                           </span>
                         </div>
 
@@ -96,7 +93,7 @@ export default function Comment() {
                         )}
                       </div>
 
-                      {/* ✅ 수정 중이면 textarea 표시, 아니면 기존 텍스트 표시 */}
+                      {/*  수정 중이면 textarea 표시, 아니면 기존 텍스트 표시 */}
                       {editingComments[comment.id] ? (
                         <div className="p-4 rounded-xl border focus-within:border-green-500 cursor-text">
                           <textarea
@@ -131,7 +128,6 @@ export default function Comment() {
                         <p className="text-gray-700">{comment.content}</p>
                       )}
 
-                      {/* 🟢 답글 달기 버튼 */}
                       {!replyingComments[comment.id] && (
                         <button
                           className="text-sm text-green-600 mt-1 hover:underline"
@@ -143,13 +139,11 @@ export default function Comment() {
                     </div>
                   </div>
 
-                  {/* 🟢 대댓글 입력창 (isReplying이 true일 때만 보이게) */}
                   {replyingComments[comment.id] && (
                     <div
                       className="ml-16 p-4 rounded-xl border focus-within:border-green-500 cursor-text"
                       ref={replyContainerRef}
                     >
-                      {/* 대댓글 입력 필드 */}
                       <textarea
                         ref={textareaRef}
                         className="resize-none w-full px-2 py-3 rounded-md border-none bg-transparent focus:outline-none overflow-hidden"
@@ -159,7 +153,6 @@ export default function Comment() {
                         rows={1}
                       ></textarea>
 
-                      {/* 버튼 영역 */}
                       <div className="flex justify-end gap-2 mt-3">
                         <button
                           className="px-4 py-2 rounded-2xl text-gray-600 bg-[#E9E8E2] transition hover:bg-gray-300"
@@ -179,7 +172,6 @@ export default function Comment() {
                     </div>
                   )}
 
-                  {/* 🟢 대댓글 목록 */}
                   <div className="flex flex-col gap-4 ml-16 space-y-3">
                     {comment.replies?.map((reply) => (
                       <div
@@ -200,13 +192,17 @@ export default function Comment() {
                                 {reply.writeUserName}
                               </span>
                               <span className="text-gray-500 text-sm">
-                                {new Date(reply.createdAt).toLocaleDateString(
-                                  "ko-KR"
-                                )}
+                                {new Date(comment.createdAt)
+                                  .toLocaleDateString("ko-KR", {
+                                    year: "numeric",
+                                    month: "2-digit",
+                                    day: "2-digit",
+                                  })
+                                  .replace(/-/g, ".")}{" "}
+                                {/* 하이픈을 점으로 변경 */}
                               </span>
                             </div>
 
-                            {/* ✅ 내가 작성한 대댓글에만 수정/삭제 버튼 표시 */}
                             {user?.name === reply.writeUserName && (
                               <div className="flex space-x-6 text-sm text-gray-600">
                                 <>
@@ -235,7 +231,6 @@ export default function Comment() {
                             )}
                           </div>
 
-                          {/* ✅ 수정 중이면 textarea 표시, 아니면 기존 텍스트 표시 */}
                           {editingReplies[reply.id] ? (
                             <div className="p-4 rounded-xl border focus-within:border-green-500 cursor-text">
                               <textarea
