+++ src/components/communityBoard/CommunityBoard-detail/Comment/type.ts
@@ -12,5 +12,5 @@ export interface CommentType {
   writeUserName: string;
   content: string;
   createdAt: string;
-  replies?: Reply[]; // ✅ replies 속성 추가 (대댓글)
+  replies?: Reply[];
 }
