+++ src/components/communityBoard/CommunityBoard-detail/Footer/hook.ts
@@ -4,15 +4,13 @@ import { useEffect, useState } from "react";
 
 export function useFooter(postId: number, fetchComments: () => void) {
   const [bookmarkToggle, setBookmarkToggle] = useState(false);
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [bookmarkCount, setBookmarkCount] = useState(0);
+  const [, setBookmarkCount] = useState(0);
   const [inputValue, setInputValue] = useState("");
 
   const token = localStorage.getItem("token-storage")
     ? JSON.parse(localStorage.getItem("token-storage")!).accessToken
     : null;
 
-  // ✅ 1️⃣ 초기 북마크 상태 불러오기
   useEffect(() => {
     const fetchBookmarkStatus = async () => {
       try {
@@ -38,9 +36,8 @@ export function useFooter(postId: number, fetchComments: () => void) {
     };
 
     fetchBookmarkStatus();
-  }, [postId, token]); // ✅ postId가 변경될 때 실행
+  }, [postId, token]);
 
-  // ✅ 2️⃣ 북마크 토글
   const toggleBookmark = async (postId: number) => {
     if (!postId) return;
     try {
@@ -55,8 +52,6 @@ export function useFooter(postId: number, fetchComments: () => void) {
       });
 
       if (!response.ok) throw new Error(`북마크 토글 실패: ${response.status}`);
-
-      console.log("✅ 북마크 토글 성공");
       setBookmarkToggle((prev) => !prev);
       setBookmarkCount((prev) => (bookmarkToggle ? prev - 1 : prev + 1));
     } catch (error) {
@@ -65,7 +60,6 @@ export function useFooter(postId: number, fetchComments: () => void) {
     }
   };
 
-  // ✅ 3️⃣ 댓글 입력 & 등록
   const onClickSubmit = async (postId: number, comment: string) => {
     if (!postId || !comment.trim()) {
       alert("댓글을 입력해주세요.");
@@ -98,8 +92,8 @@ export function useFooter(postId: number, fetchComments: () => void) {
       console.log("✅ 댓글 등록 성공");
 
       alert("댓글이 등록되었습니다!");
-      setInputValue(""); // ✅ 댓글 입력창 초기화
-      fetchComments(); // ✅ 댓글 리스트 갱신
+      setInputValue("");
+      fetchComments();
     } catch (error) {
       console.error("❌ 댓글 등록 실패:", error);
       alert("댓글 등록에 실패했습니다.");
