+++ src/components/communityBoard/CommunityBoard-list/BoardItem/index.tsx
@@ -34,7 +34,7 @@ export default function CommunityBoardItem({ post }: { post: CommunityPost }) {
               <div className="flex items-center">
                 <Image
                   className="w-full h-full"
-                  src="/icons/post_list_view_icon_24px.svg" // 조회수
+                  src="/icons/post_list_view_icon_24px.svg"
                   alt="Cancel Icon"
                   width={0} // 크기
                   height={0}
@@ -44,9 +44,9 @@ export default function CommunityBoardItem({ post }: { post: CommunityPost }) {
               <div className="flex items-center">
                 <Image
                   className="w-full h-full"
-                  src="/icons/community_detail_bookmark_24px.svg" // 저장수
+                  src="/icons/community_detail_bookmark_24px.svg"
                   alt="Cancel Icon"
-                  width={0} // 크기
+                  width={0}
                   height={0}
                 />
                 <span>{post.bookmarkCount}</span>
@@ -54,9 +54,9 @@ export default function CommunityBoardItem({ post }: { post: CommunityPost }) {
               <div className="flex items-center">
                 <Image
                   className="w-full h-full"
-                  src="/icons/post_list_chat_icon_24px.svg" // 댓글수
+                  src="/icons/post_list_chat_icon_24px.svg"
                   alt="Cancel Icon"
-                  width={0} // 크기
+                  width={0}
                   height={0}
                 />
                 <span>{post.commentCount}</span>
@@ -68,7 +68,7 @@ export default function CommunityBoardItem({ post }: { post: CommunityPost }) {
               className="w-full h-full rounded-2xl bg-center bg-cover bg-no-repeat flex-shrink-0"
               style={{
                 backgroundImage: `url(${post.thumbnailImage})`,
-                backgroundColor: "#d3d3d3", // 썸네일이 없다면?
+                backgroundColor: "#d3d3d3",
               }}
             ></div>
           </div>
